refactor: extract utility functions to lib/
- Moved get_cwd() and read_input() from Shell to lib/utils
- Added get_username() and get_hostname() to lib/utils
- Shell constructor now delegates to utility functions
