+++ src/shell.hpp
@@ -7,8 +7,8 @@
  * 
  * @author Filipe Paredes (filipeparedes3@gmail.com)
  * 
- * @version 0.1
- * @date 2026-04-28
+ * @version 0.1.1
+ * @date 2026-04-29
  * 
  * @copyright Copyright (c) 2026
  * 
@@ -53,20 +53,4 @@
          * 
          */
         void print_prompt() const;
-
-        /**
-         * @brief Reads a line of input from stdin.
-         * 
-         * @return The line entered by the user, or an empty string on EOF.
-         */
-        std::string read_input() const;
-
-        /**
-         * @brief Resolves the current working directory.
-         * 
-         * Replaces the home directory prefix with - if applicable.
-         * 
-         * @return A human-readable representation of the current directory.
-         */
-        std::string get_cwd() const;
  };
